pyproject.toml
[build-system]

requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]

name = "xy_django_serializer"

authors = [{ name = "余洋", email = "yuyangit.0515@qq.com" }]
description = "django模型解析基类"

dependencies = ["django", "djangorestframework"]

readme = "README.md"

requires-python = ">=3.11.3"

dynamic = ["version"]

classifiers = ["Programming Language :: Python :: 3"]

[project.scripts]

[tool.setuptools.packages.find]
where = ["./"]
include = ["xy_django_serializer"] # alternatively: `exclude = ["additional*"]`
namespaces = false

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { file = "VERSION" }
